+++ resources/views/admin/partner/index.blade.php
@@ -22,16 +22,19 @@
         <div class="flex-1">
             <h2 class="w-full text-2xl font-bold text-secondary ">Our Clients</h2>
         </div>
-        <div class="text-right "> <a href="{{ route('admin.partners.create') }} "
-                class="flex items-center px-3 py-2 mb-1 mr-1 text-xs font-bold text-white uppercase transition-all ease-linear bg-[#6a68AF] border border-[#6a68AF] rounded outline-none hover:bg-transparent hover:text-[#6a68AF] focus:outline-none duration-400 ">
-                <svg xmlns="http://www.w3.org/2000/svg" class="icon icon-tabler icon-tabler-plus" width="24" height="24"
-                    viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round"
-                    stroke-linejoin="round">
-                    <path stroke="none" d="M0 0h24v24H0z" fill="none"></path>
-                    <path d="M12 5l0 14"></path>
-                    <path d="M5 12l14 0"></path>
-                </svg>
-                Add Clients</a>
+        <div class="text-right ">
+            @can('Add Client')
+                <a href="{{ route('admin.partners.create') }} "
+                    class="flex items-center px-3 py-2 mb-1 mr-1 text-xs font-bold text-white uppercase transition-all ease-linear bg-[#6a68AF] border border-[#6a68AF] rounded outline-none hover:bg-transparent hover:text-[#6a68AF] focus:outline-none duration-400 ">
+                    <svg xmlns="http://www.w3.org/2000/svg" class="icon icon-tabler icon-tabler-plus" width="24" height="24"
+                        viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round"
+                        stroke-linejoin="round">
+                        <path stroke="none" d="M0 0h24v24H0z" fill="none"></path>
+                        <path d="M12 5l0 14"></path>
+                        <path d="M5 12l14 0"></path>
+                    </svg>
+                    Add Clients</a>
+            @endcan
         </div>
     </div>
 
@@ -48,10 +51,10 @@
                                 <th class="px-5 py-3 font-semibold text-left ">
                                     Title</th>
 
-                                    <th class="px-5 py-3 font-semibold text-left ">
-                                       Order</th>
-                                       <th class="px-5 py-3 font-semibold text-left ">
-                                        Image</th>
+                                <th class="px-5 py-3 font-semibold text-left ">
+                                    Order</th>
+                                <th class="px-5 py-3 font-semibold text-left ">
+                                    Image</th>
 
                                 <th class="px-5 py-3 font-semibold text-left ">
                                     Created at</th>
@@ -72,7 +75,7 @@
 
 
                                     <td class="px-5 py-3 ">
-                                        <p class="text-gray-900 ">{{ $blog->order}}</p>
+                                        <p class="text-gray-900 ">{{ $blog->order }}</p>
                                     </td>
 
 
@@ -90,47 +93,51 @@
 
                                     <td>
                                         <div class="flex items-center p-2">
+                                            @can('Edit Client')
+                                                <a href=" {{ route('admin.partners.edit', $blog->id) }}">
+                                                    <div class="flex px-2 py-1 mx-2 text-white rounded-md bg-slate-500">
+                                                        <svg xmlns="http://www.w3.org/2000/svg"
+                                                            class="icon icon-tabler icon-tabler-edit" width="24"
+                                                            height="24" viewBox="0 0 24 24" stroke-width="2"
+                                                            stroke="currentColor" fill="none" stroke-linecap="round"
+                                                            stroke-linejoin="round">
+                                                            <path stroke="none" d="M0 0h24v24H0z" fill="none"></path>
+                                                            <path
+                                                                d="M7 7h-1a2 2 0 0 0 -2 2v9a2 2 0 0 0 2 2h9a2 2 0 0 0 2 -2v-1">
+                                                            </path>
+                                                            <path
+                                                                d="M20.385 6.585a2.1 2.1 0 0 0 -2.97 -2.97l-8.415 8.385v3h3l8.385 -8.415z">
+                                                            </path>
+                                                            <path d="M16 5l3 3"></path>
+                                                        </svg>
 
-                                            <a href=" {{ route('admin.partners.edit', $blog->id) }}">
-                                                <div class="flex px-2 py-1 mx-2 text-white rounded-md bg-slate-500">
-                                                    <svg xmlns="http://www.w3.org/2000/svg"
-                                                        class="icon icon-tabler icon-tabler-edit" width="24"
-                                                        height="24" viewBox="0 0 24 24" stroke-width="2"
-                                                        stroke="currentColor" fill="none" stroke-linecap="round"
-                                                        stroke-linejoin="round">
-                                                        <path stroke="none" d="M0 0h24v24H0z" fill="none"></path>
-                                                        <path
-                                                            d="M7 7h-1a2 2 0 0 0 -2 2v9a2 2 0 0 0 2 2h9a2 2 0 0 0 2 -2v-1">
-                                                        </path>
-                                                        <path
-                                                            d="M20.385 6.585a2.1 2.1 0 0 0 -2.97 -2.97l-8.415 8.385v3h3l8.385 -8.415z">
-                                                        </path>
-                                                        <path d="M16 5l3 3"></path>
-                                                    </svg>
+                                                    </div>
+                                                </a>
+                                            @endcan
 
-                                                </div>
-                                            </a>
 
-                                            <form method="POST" action="{{ route('admin.partners.destroy', $blog->id) }}"
-                                                id="delete-form-{{ $blog->id }}">
-                                                @csrf
-                                                @method('delete')
-                                                <button type="button" onclick="deleteItem({{ $blog->id }})"
-                                                    class="flex px-2 py-1 mx-2 text-white bg-red-500 rounded-md">
-                                                    <svg xmlns="http://www.w3.org/2000/svg"
-                                                        class="icon icon-tabler icon-tabler-trash" width="24"
-                                                        height="24" viewBox="0 0 24 24" stroke-width="2"
-                                                        stroke="currentColor" fill="none" stroke-linecap="round"
-                                                        stroke-linejoin="round">
-                                                        <path stroke="none" d="M0 0h24v24H0z" fill="none"></path>
-                                                        <path d="M4 7l16 0"></path>
-                                                        <path d="M10 11l0 6"></path>
-                                                        <path d="M14 11l0 6"></path>
-                                                        <path d="M5 7l1 12a2 2 0 0 0 2 2h8a2 2 0 0 0 2 -2l1 -12"></path>
-                                                        <path d="M9 7v-3a1 1 0 0 1 1 -1h4a1 1 0 0 1 1 1v3"></path>
-                                                    </svg>
-                                                </button>
-                                            </form>
+                                            @can('Delete Client')
+                                                <form method="POST" action="{{ route('admin.partners.destroy', $blog->id) }}"
+                                                    id="delete-form-{{ $blog->id }}">
+                                                    @csrf
+                                                    @method('delete')
+                                                    <button type="button" onclick="deleteItem({{ $blog->id }})"
+                                                        class="flex px-2 py-1 mx-2 text-white bg-red-500 rounded-md">
+                                                        <svg xmlns="http://www.w3.org/2000/svg"
+                                                            class="icon icon-tabler icon-tabler-trash" width="24"
+                                                            height="24" viewBox="0 0 24 24" stroke-width="2"
+                                                            stroke="currentColor" fill="none" stroke-linecap="round"
+                                                            stroke-linejoin="round">
+                                                            <path stroke="none" d="M0 0h24v24H0z" fill="none"></path>
+                                                            <path d="M4 7l16 0"></path>
+                                                            <path d="M10 11l0 6"></path>
+                                                            <path d="M14 11l0 6"></path>
+                                                            <path d="M5 7l1 12a2 2 0 0 0 2 2h8a2 2 0 0 0 2 -2l1 -12"></path>
+                                                            <path d="M9 7v-3a1 1 0 0 1 1 -1h4a1 1 0 0 1 1 1v3"></path>
+                                                        </svg>
+                                                    </button>
+                                                </form>
+                                            @endcan
 
                                         </div>
                                     </td>
